Don't get stuck for closed-shell molecules
The previous code would wait until the question of whether or not to
write out natural orbitals before assuming that we have finished the
occupation assignment. However, this question is never asked for
closed-shell molecules causing a timeout error.

Therefore, we now (more cleverly) simply detect whether the new menu has
opened and if it has, we assume that the occupation assignment is over.